requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.20
Werkzeug==2.3.7
Flask-Migrate==4.0.4
email-validator==2.0.0
Flask-Bootstrap4==4.0.2 